backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
langgraph>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.1

